requirements.txt
# requirements.txt
python-fasthtml
monsterui
python-dotenv
google-generativeai
pdfplumber
pillow
markdown
python-docx

